train/requirements-server.txt
# Requirements for model training and serving
fastapi>=0.104,<1
uvicorn>=0.24,<1
numpy>=1.24.0
tensorflow>=2.13.0
keras>=2.13.0
scikit-learn>=1.3.0
joblib>=1.3.0
Pillow>=10.0.0
